Extract DarkModePanel and PieceSymbolPanel as widgets.
