+++ lib/widgets/home_screen/settings_tab/settings_tab.dart
@@ -1,17 +1,10 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:shogi_proverbs/di_container.dart';
 import 'package:shogi_proverbs/localizations.dart';
-import 'package:shogi_proverbs/services/settings_database/i_settings_database.dart';
-import 'package:shogi_proverbs/widgets/home_screen/settings_tab/segmented_chips.dart';
-
-final isDarkModeProvider = StateProvider((_) => DIContainer.get<ISettingsDatabase>().isDarkMode);
-final selectedPieceSymbolProvider = StateProvider((_) => DIContainer.get<ISettingsDatabase>().selectedPieceSymbol);
+import 'package:shogi_proverbs/widgets/common/panels/dark_mode_panel.dart';
+import 'package:shogi_proverbs/widgets/common/panels/piece_symbol_panel.dart';
 
 class SettingsTab extends StatelessWidget {
-  static const _languageSymbols = ['K', '玉'];
-
   const SettingsTab({Key key}) : super(key: key);
 
   @override
@@ -24,35 +17,9 @@ class SettingsTab extends StatelessWidget {
             crossAxisAlignment: CrossAxisAlignment.start,
             children: <Widget>[
               SizedBox(height: 16.0),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: <Widget>[
-                  Text(AppLocalizations.settingsTabDarkModeLabel),
-                  Consumer((_, read) => Switch(
-                        value: read(isDarkModeProvider).state,
-                        onChanged: (value) {
-                          isDarkModeProvider.read(context).state = value;
-                          DIContainer.get<ISettingsDatabase>().isDarkMode = value;
-                        },
-                      )),
-                ],
-              ),
+              DarkModePanel(),
               SizedBox(height: 16.0),
-              Row(
-                crossAxisAlignment: CrossAxisAlignment.center,
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Text(AppLocalizations.settingsTabPieceSymbolLabel),
-                  Consumer((_, read) => SegmentedChips(
-                        labels: _languageSymbols,
-                        initiallySelectedIndex: read(selectedPieceSymbolProvider).state,
-                        onSelected: (selectedIndex) {
-                          selectedPieceSymbolProvider.read(context).state = selectedIndex;
-                          DIContainer.get<ISettingsDatabase>().selectedPieceSymbol = selectedIndex;
-                        },
-                      )),
-                ],
-              ),
+              PieceSymbolPanel(),
               SizedBox(height: 16.0),
               Center(
                 child: RaisedButton(
