+++ lib/widgets/my_app.dart
@@ -7,8 +7,9 @@ import 'package:shogi_proverbs/configs/route_names.dart';
 import 'package:shogi_proverbs/di_container.dart';
 import 'package:shogi_proverbs/localizations.dart';
 import 'package:shogi_proverbs/services/settings_database/i_settings_database.dart';
+import 'package:shogi_proverbs/widgets/common/panels/dark_mode_panel.dart';
+import 'package:shogi_proverbs/widgets/common/panels/piece_symbol_panel.dart';
 import 'package:shogi_proverbs/widgets/home_screen/home_screen.dart';
-import 'package:shogi_proverbs/widgets/home_screen/settings_tab/settings_tab.dart';
 import 'package:shogi_proverbs/widgets/onboarding/onboarding_screen.dart';
 
 class MyApp extends StatefulWidget {
